refactor(users.php): Move user management functions
Put the `addUser()`, `deleteUser()`, `changeUserPassword()`, and
`getUsers()` functions into their own file, out of misc/functions.php.
