+++ PodcastGenerator/core/include_admin.php
@@ -36,6 +36,7 @@ include 'freebox.php';
 // This file is wizard to convert old password to a more secure algorithm
 // Load useful functions
 include 'misc/functions.php';
+include 'users.php';
 // Load HTML helper functions
 include 'html_helpers.php';
 // Load translations
